+++ modules/tinymce/src/themes/silver/main/ts/api/Settings.ts
@@ -183,8 +183,6 @@ const getFilePickerValidatorHandler = (editor: Editor) => {
   }
 };
 
-const getIframeAttrs = (editor: Editor): Record<string, string> => editor.getParam('iframe_attrs', {});
-
 export {
   getSkinUrl,
   isReadOnly,
@@ -223,6 +221,5 @@ export {
   noTypeaheadUrls,
   getAnchorTop,
   getAnchorBottom,
-  getFilePickerValidatorHandler,
-  getIframeAttrs
+  getFilePickerValidatorHandler
 };